requirements.txt
absl-py==0.13.0
alabaster==0.7.12
alembic==1.7.5
altair==4.1.0
anaconda-client==1.7.2
anaconda-navigator==2.0.4
anaconda-project==0.8.3
anyio==3.4.0
apache-airflow==2.2.2
apache-airflow-providers-amazon==2.4.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-postgres==2.3.0
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
applaunchservices==0.2.1
appnope==0.1.0
appscript==1.1.1
argcomplete==1.12.3
argh==0.26.2
argon2-cffi==20.1.0
asn1crypto==1.4.0
astor==0.8.1
astroid==2.5
astropy==4.0.2
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
Babel==2.8.1
backcall==0.2.0
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
base58==2.1.0
beautifulsoup4==4.9.3
bitarray==1.6.1
bkcharts==0.2
bleach==3.2.1
blinker==1.4
bokeh==2.2.3
boto==2.49.0
boto3==1.18.65
botocore==1.23.16
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
cattrs==1.5.0
certifi==2021.5.30
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.0.8
clang==5.0
click==7.1.2
clickclick==20.10.2
cloudpickle==1.6.0
clyent==1.2.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
conda==4.10.3
conda-build==3.20.5
conda-content-trust==0+unknown
conda-package-handling==1.7.2
conda-repo-cli==1.0.4
conda-token==0.3.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
croniter==1.0.15
cryptography==3.1.1
cycler==0.10.0
Cython==0.29.21
cytoolz==0.11.0
dask==2.30.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dill==0.3.4
distributed==2.30.1
dnspython==2.1.0
docopt==0.6.2
docutils==0.16
email-validator==1.1.3
entrypoints==0.3
et-xmlfile==1.0.1
fastapi==0.68.1
fastcache==1.1.0
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-AppBuilder==3.4.0
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flatbuffers==1.12
fsspec==0.8.3
future==0.18.2
gast==0.4.0
gevent==20.9.0
gitdb==4.0.7
GitPython==3.1.18
glob2==0.7
gmpy2==2.0.8
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.18.2
greenlet==0.4.17
grpcio==1.40.0
gunicorn==20.1.0
h11==0.12.0
h2o==3.32.1.6
h5py==3.1.0
HeapDict==1.0.1
html5lib==1.1
httpcore==0.14.3
httpx==0.21.1
hyperopt==0.2.5
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==2.0.0
importlib-resources==5.4.0
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==1.0.2
isodate==0.6.0
isort==5.6.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json5==0.9.5
jsonpath-ng==1.5.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keras==2.6.0
Keras-Preprocessing==1.1.2
keyring==21.4.0
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.34.0
locket==0.2.0
lockfile==0.12.2
lxml==4.6.4
Mako==1.1.6
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.6.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
navigator-updater==0.2.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
networkx==2.5
nltk==3.5
nose==1.3.7
notebook==6.1.4
numba==0.51.2
numexpr==2.7.1
numpy==1.19.2
numpydoc==1.1.0
oauthlib==3.1.1
olefile==0.46
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.0
partd==1.1.0
path==15.0.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pendulum==2.1.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip==20.2.4
pipreqs==0.4.10
pkginfo==1.6.1
plotly==5.1.0
pluggy==0.13.1
ply==3.11
prison==0.2.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.17.3
psutil==5.7.2
psycopg2==2.9.2
psycopg2-binary==2.9.2
ptyprocess==0.6.0
py==1.9.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycurl==7.43.0.6
pydantic==1.8.2
pydeck==0.6.2
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
pyodbc==4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyprojroot==0.2.0
pyrsistent==0.17.3
PySocks==1.7.1
pytest==0.0.0
python-daemon==2.3.0
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2020.1
pytzdata==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtAwesome==1.0.1
qtconsole==4.7.7
QtPy==1.9.0
redshift-connector==2.0.900
regex==2020.10.15
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==10.14.0
rope==0.18.0
rsa==4.7.2
Rtree==0.9.4
ruamel-yaml==0.15.87
s3transfer==0.5.0
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
scramp==1.4.1
seaborn==0.11.0
Send2Trash==1.5.0
setproctitle==1.2.2
setuptools==50.3.1.post20201107
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
smmap==4.0.0
sniffio==1.2.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
spyder==4.1.5
spyder-kernels==1.9.4
SQLAlchemy==1.3.20
SQLAlchemy-JSONField==1.0.0
sqlalchemy-redshift==0.8.8
SQLAlchemy-Utils==0.37.9
starlette==0.14.2
statsmodels==0.12.0
streamlit==0.84.2
swagger-ui-bundle==0.0.9
sympy==1.6.2
tables==3.6.1
tabulate==0.8.9
tblib==1.7.0
tenacity==8.0.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2020.10.1
toml==0.10.1
toolz==0.11.1
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchvision==0.8.2
tornado==6.0.4
tqdm==4.50.2
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==4.0.1
unicodecsv==0.14.1
urllib3==1.25.11
validators==0.18.2
watchdog==0.10.3
watchtower==1.0.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.35.1
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
wurlitzer==2.0.1
xgboost==1.4.2
xlrd==1.2.0
XlsxWriter==1.3.7
xlwings==0.20.8
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.30.0
yarg==0.1.9
zict==2.0.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
